+++ consul-tls/consul-tls.sh
@@ -97,6 +97,9 @@ pkg install -y curl
 step "Install package jq"
 pkg install -y jq
 
+step "Install package syslog-ng"
+pkg install -y syslog-ng
+
 step "Install package vault"
 # removed - using ports to get 1.7.3
 # we install vault to get the correct rc files
@@ -245,20 +248,7 @@ fi
 
 # setup directories for vault usage
 mkdir -p /mnt/templates
-mkdir -p /mnt/certs
-
-# optional remote logging
-# to-do: evaluate valid ip/domain name
-mkdir -p /usr/local/etc/syslog.d
-if [ ! -z \$REMOTELOG ] && [ \$REMOTELOG != \"null\" ]; then
-    echo \"# send logs to remote syslog loki instance
-*.*        @\$REMOTELOG:514
-\" > /usr/local/etc/syslog.d/remotelog.conf
-    /etc/rc.d/syslogd restart
-else
-    echo \"REMOTELOG parameter is not set to an IP address\"
-fi
-
+mkdir -p /mnt/certs/localca
 
 ## start Vault
 
@@ -288,15 +278,15 @@ storage \\\"file\\\" {
 }
 template {
   source = \\\"/mnt/templates/cert.tpl\\\"
-  destination = \\\"/mnt/certs/nomadcert.pem\\\"
+  destination = \\\"/mnt/certs/consulcert.pem\\\"
 }
 template {
   source = \\\"/mnt/templates/ca.tpl\\\"
-  destination = \\\"/mnt/certs/nomadca.pem\\\"
+  destination = \\\"/mnt/certs/consulca.pem\\\"
 }
 template {
   source = \\\"/mnt/templates/key.tpl\\\"
-  destination = \\\"/mnt/certs/nomadkey.pem\\\"
+  destination = \\\"/mnt/certs/consulkey.pem\\\"
 }\" > /usr/local/etc/vault.hcl
 
 # setup template files for certificates
@@ -362,6 +352,8 @@ if [ -s /root/login.token ]; then
     #/usr/local/bin/jq -r '.data.issuing_ca[]' /mnt/certs/vaultissue.json > /mnt/certs/nomadca.pem
     # if [] left in for this script, you will get error: Cannot iterate over string
     /usr/local/bin/jq -r '.data.issuing_ca' /mnt/certs/vaultissue.json > /mnt/certs/consulca.pem
+    # syslog-ng wants ca file in a directory, so copy CA file to there too - not currently in use
+    cp -f /mnt/certs/consulca.pem /mnt/certs/localca/consulca.pem
     /usr/local/bin/jq -r '.data.certificate' /mnt/certs/vaultissue.json > /mnt/certs/consulcert.pem
     /usr/local/bin/jq -r '.data.private_key' /mnt/certs/vaultissue.json > /mnt/certs/consulkey.pem
 
@@ -380,12 +372,15 @@ if [ -s /root/login.token ]; then
     HEADER=\\\$(echo \\\"X-Vault-Token: \\\"\\\$LOGINTOKEN)
     /usr/local/bin/curl -k --header \\\"\\\$HEADER\\\" --request POST --data @/mnt/templates/payload.json https://\$VAULTSERVER:8200/v1/pki_int/issue/\$DATACENTER > /mnt/certs/vaultissue.json
     /usr/local/bin/jq -r '.data.issuing_ca' /mnt/certs/vaultissue.json > /mnt/certs/consulca.pem
+    # syslog-ng wants ca file in a directory, so copy CA file to there too - not currently in use
+    cp -f /mnt/certs/consulca.pem /mnt/certs/localca/consulca.pem
     /usr/local/bin/jq -r '.data.certificate' /mnt/certs/vaultissue.json > /mnt/certs/consulcert.pem
     /usr/local/bin/jq -r '.data.private_key' /mnt/certs/vaultissue.json > /mnt/certs/consulkey.pem
     # set permissions on /mnt/certs for vault
     chown -R vault:wheel /mnt/certs
     #/bin/pkill -HUP consul
-    /usr/local/etc/rc.d/consul reload
+    /usr/local/etc/rc.d/consul restart
+    /usr/local/etc/rc.d/syslog-ng restart
 else
     echo "/root/login.token does not contain a token. Certificates cannot be renewed."
 fi
@@ -400,6 +395,28 @@ fi
 
 else
     echo \"ERROR: There was a problem logging into vault and no certificates were retrieved. Vault not started.\"
+fi
+
+# setup syslog-ng
+# optional remote logging
+if [ ! -z \$REMOTELOG ] && [ \$REMOTELOG != \"null\" ]; then
+    if [ -f /root/syslog-ng.conf ]; then
+        /usr/bin/sed -i .orig \"s/REMOTELOGIP/\$REMOTELOG/g\" /root/syslog-ng.conf
+        cp -f /root/syslog-ng.conf /usr/local/etc/syslog-ng.conf
+        # stop syslogd
+        service syslogd onestop || true
+        # setup sysrc entries to start and set parameters to accept logs from remote subnet
+        sysrc syslogd_enable=\"NO\"
+        sysrc syslog_ng_enable=\"YES\"
+        #sysrc syslog_ng_flags=\"-u daemon\"
+        sysrc syslog_ng_flags=\"-R /tmp/syslog-ng.persist\"
+        /usr/local/etc/rc.d/syslog-ng start
+        echo \"syslog-ng setup complete\"
+    else
+        echo \"/root/syslog-ng.conf is missing?\"
+    fi
+else
+    echo \"REMOTELOG parameter is not set to an IP address. syslog-ng won't operate.\"
 fi
 
 # Create consul server config file, set the bootstrap_expect value to number