+++ postgresql-patroni/postgresql-patroni.sh
@@ -96,6 +96,9 @@ pkg install -y consul
 step "Install package node_exporter"
 pkg install -y node_exporter
 
+step "Install package syslog-ng"
+pkg install -y syslog-ng
+
 step "Install package postgresql-server"
 pkg install -y postgresql13-server
 
@@ -266,18 +269,7 @@ fi
 
 # setup directories for vault usage
 mkdir -p /mnt/templates
-mkdir -p /mnt/certs
-
-# optional remote logging
-mkdir -p /usr/local/etc/syslog.d
-if [ ! -z \$REMOTELOG ] && [ \$REMOTELOG != \"null\" ]; then
-    echo \"# send logs to remote syslog loki instance
-*.*        @\$REMOTELOG:514
-\" > /usr/local/etc/syslog.d/remotelog.conf
-    /etc/rc.d/syslogd restart
-else
-    echo \"REMOTELOG parameter is not set to an IP address\"
-fi
+mkdir -p /mnt/certs/localca
 
 # make consul configuration directory and set permissions
 mkdir -p /usr/local/etc/consul.d
@@ -447,6 +439,8 @@ if [ -s /root/login.token ]; then
     #/usr/local/bin/jq -r '.data.issuing_ca[]' /mnt/certs/vaultissue.json > /mnt/certs/postgresca.pem
     # if [] left in for this script, you will get error: Cannot iterate over string
     /usr/local/bin/jq -r '.data.issuing_ca' /mnt/certs/vaultissue.json > /mnt/certs/postgresca.pem
+    # syslog-ng wants ca file in a directory, so copy CA file to there too - not currently in use
+    cp -f /mnt/certs/postgresca.pem /mnt/certs/localca/postgresca.pem
     /usr/local/bin/jq -r '.data.certificate' /mnt/certs/vaultissue.json > /mnt/certs/postgrescert.pem
     /usr/local/bin/jq -r '.data.private_key' /mnt/certs/vaultissue.json > /mnt/certs/postgreskey.pem
 
@@ -468,11 +462,13 @@ if [ -s /root/login.token ]; then
     HEADER=\\\$(echo \\\"X-Vault-Token: \\\"\\\$LOGINTOKEN)
     /usr/local/bin/curl -k --header \\\"\\\$HEADER\\\" --request POST --data @/mnt/templates/payload.json https://\$VAULTSERVER:8200/v1/pki_int/issue/\$DATACENTER > /mnt/certs/vaultissue.json
     /usr/local/bin/jq -r '.data.issuing_ca' /mnt/certs/vaultissue.json > /mnt/certs/postgresca.pem
+    # syslog-ng wants ca file in a directory, so copy CA file to there too - not currently in use
+    cp -f /mnt/certs/postgresca.pem /mnt/certs/localca/postgresca.pem
     /usr/local/bin/jq -r '.data.certificate' /mnt/certs/vaultissue.json > /mnt/certs/postgrescert.pem
     /usr/local/bin/jq -r '.data.private_key' /mnt/certs/vaultissue.json > /mnt/certs/postgreskey.pem
     # set permissions on /mnt/certs for vault
     chown -R vault:wheel /mnt/certs
-    /usr/local/etc/rc.d/consul reload
+    /usr/local/etc/rc.d/consul restart
     /usr/local/etc/rc.d/node_exporter restart
     # restart gives error with port already in use
     #/usr/local/etc/rc.d/patroni restart
@@ -488,6 +484,29 @@ fi
         # add a crontab entry for every hour
         echo \"0 * * * * root /root/rotate-certs.sh >> /mnt/rotate-cert.log 2>&1\" >> /etc/crontab
     fi
+
+    # setup syslog-ng
+    # optional remote logging
+    if [ ! -z \$REMOTELOG ] && [ \$REMOTELOG != \"null\" ]; then
+        if [ -f /root/syslog-ng.conf ]; then
+            /usr/bin/sed -i .orig \"s/REMOTELOGIP/\$REMOTELOG/g\" /root/syslog-ng.conf
+            cp -f /root/syslog-ng.conf /usr/local/etc/syslog-ng.conf
+            # stop syslogd
+            service syslogd onestop || true
+            # setup sysrc entries to start and set parameters to accept logs from remote subnet
+            sysrc syslogd_enable=\"NO\"
+            sysrc syslog_ng_enable=\"YES\"
+            #sysrc syslog_ng_flags=\"-u daemon\"
+            sysrc syslog_ng_flags=\"-R /tmp/syslog-ng.persist\"
+            /usr/local/etc/rc.d/syslog-ng start
+            echo \"syslog-ng setup complete\"
+        else
+            echo \"/root/syslog-ng.conf is missing?\"
+        fi
+    else
+        echo \"REMOTELOG parameter is not set to an IP address. syslog-ng won't operate.\"
+    fi
+
     ## start node_exporter config
     # node exporter needs tls setup
     echo \"tls_server_config: