+++ vault/vault.sh
@@ -100,6 +100,9 @@ pkg install -y jq
 step "Install package curl"
 pkg install -y curl
 
+step "Install package syslog-ng"
+pkg install -y syslog-ng
+
 step "Clean package installation"
 pkg clean -y
 
@@ -174,8 +177,10 @@ then
 fi
 
 # ADJUST THIS: STOP SERVICES AS NEEDED BEFORE CONFIGURATION
-/usr/local/etc/rc.d/consul onestop || true
-/usr/local/etc/rc.d/vault onestop  || true
+
+# removing as not needed, services aren't running
+#/usr/local/etc/rc.d/consul onestop || true
+#/usr/local/etc/rc.d/vault onestop  || true
 
 # No need to adjust this:
 # If this pot flavour is not blocking, we need to read the environment first from /tmp/environment.sh
@@ -258,18 +263,7 @@ fi
 
 # setup directories for vault usage
 mkdir -p /mnt/templates
-mkdir -p /mnt/certs
-
-# optional remote logging
-mkdir -p /usr/local/etc/syslog.d
-if [ ! -z \$REMOTELOG ] && [ \$REMOTELOG != \"null\" ]; then
-    echo \"# send logs to remote syslog loki instance
-*.*        @\$REMOTELOG:514
-\" > /usr/local/etc/syslog.d/remotelog.conf
-    /etc/rc.d/syslogd restart
-else
-    echo \"REMOTELOG parameter is not set to an IP address\"
-fi
+mkdir -p /mnt/certs/localca
 
 ## start Vault
 
@@ -601,6 +595,8 @@ template {
         HEADER=\$(/bin/cat /root/login.token)
         /usr/local/bin/curl --header \"X-Vault-Token: \$HEADER\" --request POST --data @/mnt/templates/payload.json http://\$IP:8200/v1/pki_int/issue/\$DATACENTER > /mnt/certs/vaultissue.json
         /usr/local/bin/jq -r '.data.issuing_ca' /mnt/certs/vaultissue.json > /mnt/certs/vaultca.pem
+        # syslog-ng wants ca file in a directory, so copy CA file to there too - not currently in use
+        cp -f /mnt/certs/vaultca.pem /mnt/certs/localca/vaultca.pem
         /usr/local/bin/jq -r '.data.certificate' /mnt/certs/vaultissue.json > /mnt/certs/vaultcert.pem
         /usr/local/bin/jq -r '.data.private_key' /mnt/certs/vaultissue.json > /mnt/certs/vaultkey.pem
         # set permissions on /mnt/certs for vault
@@ -622,6 +618,27 @@ template {
 
         # enable consul components
         /usr/bin/sed -i .orig 's/#brb#//g' /usr/local/etc/vault.hcl
+
+        # optional remote logging
+        if [ ! -z \$REMOTELOG ] && [ \$REMOTELOG != \"null\" ]; then
+            if [ -f /root/syslog-ng.conf ]; then
+                /usr/bin/sed -i .orig \"s/REMOTELOGIP/\$REMOTELOG/g\" /root/syslog-ng.conf
+                cp -f /root/syslog-ng.conf /usr/local/etc/syslog-ng.conf
+                # stop syslogd
+                service syslogd onestop || true
+                # setup sysrc entries to start and set parameters to accept logs from remote subnet
+                sysrc syslogd_enable=\"NO\"
+                sysrc syslog_ng_enable=\"YES\"
+                #sysrc syslog_ng_flags=\"-u daemon\"
+                sysrc syslog_ng_flags=\"-R /tmp/syslog-ng.persist\"
+                /usr/local/etc/rc.d/syslog-ng start
+                echo \"syslog-ng setup complete\"
+            else
+                echo \"/root/syslog-ng.conf is missing?\"
+            fi
+        else
+            echo \"REMOTELOG parameter is not set to an IP address. syslog-ng won't operate.\"
+        fi
 
         ## start consul config
         # make consul configuration directory and set permissions
@@ -725,12 +742,15 @@ if [ -s /root/login.token ]; then
     HEADER=\\\$(echo \\\"X-Vault-Token: \\\"\\\$LOGINTOKEN)
     /usr/local/bin/curl -k --header \\\"\\\$HEADER\\\" --request POST --data @/mnt/templates/payload.json https://\$VAULTLEADER:8200/v1/pki_int/issue/\$DATACENTER > /mnt/certs/vaultissue.json
     /usr/local/bin/jq -r '.data.issuing_ca' /mnt/certs/vaultissue.json > /mnt/certs/vaultca.pem
+    # syslog-ng wants ca file in a directory, so copy CA file to there too - not currently in use
+    cp -f /mnt/certs/vaultca.pem /mnt/certs/localca/vaultca.pem
     /usr/local/bin/jq -r '.data.certificate' /mnt/certs/vaultissue.json > /mnt/certs/vaultcert.pem
     /usr/local/bin/jq -r '.data.private_key' /mnt/certs/vaultissue.json > /mnt/certs/vaultkey.pem
     # set permissions on /mnt/certs for vault
     chown -R vault:wheel /mnt/certs
     /bin/pkill -HUP vault
-    /usr/local/bin/consul reload
+    /usr/local/etc/rc.d/syslog-ng restart
+    /usr/local/etc/rc.d/consul restart
 else
     echo "/root/login.token does not contain a token. Certificates cannot be renewed."
 fi
@@ -908,6 +928,8 @@ cluster_addr = \\\"https://\$IP:8201\\\"
         #/usr/local/bin/jq -r '.data.issuing_ca[]' /mnt/certs/vaultissue.json > /mnt/certs/vaultca.pem
         # if [] left in for this script, you will get error: Cannot iterate over string
         /usr/local/bin/jq -r '.data.issuing_ca' /mnt/certs/vaultissue.json > /mnt/certs/vaultca.pem
+        # syslog-ng wants ca file in a directory, so copy CA file to there too - not currently in use
+        cp -f /mnt/certs/vaultca.pem /mnt/certs/localca/vaultca.pem
         /usr/local/bin/jq -r '.data.certificate' /mnt/certs/vaultissue.json > /mnt/certs/vaultcert.pem
         /usr/local/bin/jq -r '.data.private_key' /mnt/certs/vaultissue.json > /mnt/certs/vaultkey.pem
 
@@ -916,6 +938,27 @@ cluster_addr = \\\"https://\$IP:8201\\\"
 
         # enable consul components
         /usr/bin/sed -i .orig 's/#brb#//g' /usr/local/etc/vault.hcl
+
+        # optional remote logging
+        if [ ! -z \$REMOTELOG ] && [ \$REMOTELOG != \"null\" ]; then
+            if [ -f /root/syslog-ng.conf ]; then
+                /usr/bin/sed -i .orig \"s/REMOTELOGIP/\$REMOTELOG/g\" /root/syslog-ng.conf
+                cp -f /root/syslog-ng.conf /usr/local/etc/syslog-ng.conf
+                # stop syslogd
+                service syslogd onestop || true
+                # setup sysrc entries to start and set parameters to accept logs from remote subnet
+                sysrc syslogd_enable=\"NO\"
+                sysrc syslog_ng_enable=\"YES\"
+                #sysrc syslog_ng_flags=\"-u daemon\"
+                sysrc syslog_ng_flags=\"-R /tmp/syslog-ng.persist\"
+                /usr/local/etc/rc.d/syslog-ng start
+                echo \"syslog-ng setup complete\"
+            else
+                echo \"/root/syslog-ng.conf is missing?\"
+            fi
+        else
+            echo \"REMOTELOG parameter is not set to an IP address. syslog-ng won't operate.\"
+        fi
 
         ## start consul config
         # make consul configuration directory and set permissions
@@ -1031,12 +1074,15 @@ if [ -s /root/login.token ]; then
     HEADER=\\\$(echo \\\"X-Vault-Token: \\\"\\\$LOGINTOKEN)
     /usr/local/bin/curl -k --header \\\"\\\$HEADER\\\" --request POST --data @/mnt/templates/payload.json https://\$VAULTLEADER:8200/v1/pki_int/issue/\$DATACENTER > /mnt/certs/vaultissue.json
     /usr/local/bin/jq -r '.data.issuing_ca' /mnt/certs/vaultissue.json > /mnt/certs/vaultca.pem
+    # syslog-ng wants ca file in a directory, so copy CA file to there too - not currently in use
+    cp -f /mnt/certs/vaultca.pem /mnt/certs/localca/vaultca.pem
     /usr/local/bin/jq -r '.data.certificate' /mnt/certs/vaultissue.json > /mnt/certs/vaultcert.pem
     /usr/local/bin/jq -r '.data.private_key' /mnt/certs/vaultissue.json > /mnt/certs/vaultkey.pem
     # set permissions on /mnt/certs for vault
     chown -R vault:wheel /mnt/certs
     /bin/pkill -HUP vault
-    /usr/local/bin/consul reload
+    /usr/local/etc/rc.d/syslog-ng restart
+    /usr/local/etc/rc.d/consul restart
 else
     echo "/root/login.token does not contain a token. Certificates cannot be renewed."
 fi