+++ influxdb/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.0.2
+
+* Implementing syslog-ng with tls for remote logging
+
+---
+
 0.0.1
 
 * First bash at influxdb